+++ PiCamera-client/src/app/pages/main/main.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { MatSidenav } from '@angular/material/sidenav';
 import { AdviceService } from 'src/app/services/advice/advice.service';
 import { AuthenticationService } from 'src/app/services/authentication/authentication.service';
+import { AlertData } from 'src/app/components/alert/alert.component';
 
 @Component({
   selector: 'app-main',
@@ -20,7 +21,7 @@ export class MainComponent implements OnInit {
   public ngOnInit(): void {
     this.pageList = [
       { name: 'Inicio', icon: 'home', action: () => this.navigateTo('/home') },
-      { name: 'Album', icon: 'folder', action: () => this.navigateTo('/album') },
+      { name: 'Álbum', icon: 'folder', action: () => this.navigateTo('/album') },
       { name: 'Ajustes', icon: 'settings', action: () => this.navigateTo('/settings') },
       { name: 'Cerrar sesión', icon: 'directions_run', action: () => this.closeSession() }
     ];
@@ -32,8 +33,24 @@ export class MainComponent implements OnInit {
   }
 
   public closeSession(): void {
-    this.authService.removeToken();
-    this.adviceService.showToast('La sesión se ha cerrado correctamente');
+    const options: AlertData = {
+      header: '¿Quieres cerrar la sesión?',
+      message: 'La próxima vez que accedas deberás introducir tu usuario y contraseña',
+      buttons: [
+        { name: 'Cancelar', value: 'cancel' },
+        { name: 'Aceptar', value: 'ok', isPrimary: true },
+      ]
+    };
+
+    const result = this.adviceService.showAlert(options);
+    result.subscribe(data => {
+      if (data == null || data == 'cancel') {
+        return;
+      }
+            
+      this.authService.removeToken();
+      this.adviceService.showToast('La sesión se ha cerrado correctamente');
+    });
   }
 
 }
